requirements.txt
ollama>=0.4
openai_whisper
yt-dlp>=2024.3.10
ffmpeg-python>=0.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
torch>=2.2.0
dataclasses>=0.6
